+++ datafusion/src/physical_plan/windows/built_in.rs
@@ -15,14 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
+//! Physical exec for built-in window function expressions.
+
 use crate::error::{DataFusionError, Result};
-use crate::logical_plan::window_frames::{WindowFrame, WindowFrameUnits};
+use crate::logical_plan::window_frames::WindowFrame;
 use crate::physical_plan::{
     expressions::PhysicalSortExpr,
-    window_functions::{
-        signature_for_built_in, BuiltInWindowFunction, BuiltInWindowFunctionExpr,
-        WindowFunction,
-    },
+    window_functions::{BuiltInWindowFunction, BuiltInWindowFunctionExpr},
     PhysicalExpr, WindowExpr,
 };
 use arrow::compute::concat;
@@ -39,6 +38,25 @@ pub struct BuiltInWindowExpr {
     partition_by: Vec<Arc<dyn PhysicalExpr>>,
     order_by: Vec<PhysicalSortExpr>,
     window_frame: Option<WindowFrame>,
+}
+
+impl BuiltInWindowExpr {
+    /// create a new built-in window function expression
+    pub(super) fn new(
+        fun: BuiltInWindowFunction,
+        expr: Arc<dyn BuiltInWindowFunctionExpr>,
+        partition_by: &[Arc<dyn PhysicalExpr>],
+        order_by: &[PhysicalSortExpr],
+        window_frame: Option<WindowFrame>,
+    ) -> Self {
+        Self {
+            fun,
+            expr,
+            partition_by: partition_by.to_vec(),
+            order_by: order_by.to_vec(),
+            window_frame,
+        }
+    }
 }
 
 impl WindowExpr for BuiltInWindowExpr {
